Use BRCA demo data
The existing data/TCGA.GBM.Demo.rda file has the following problem:
because the matrix data was obtained by saving the matrix directly from
an existing NGCHM in a web browser, some of the data is summary data.
This is evident by zooming in on an NGCHM made from this data: in some
regions every 4 rows have the same value.

This commit adds new BRCA demo data that was downloaded directly from
the MBatch Omic browser, and processed to create a reasonably sized
dataset. The README in this commit provides a brief explanation of the
processing.